requirements.txt
httpx>=0.24.0
fastmcp>=0.1.0
nba_api>=1.2.0
jupyter>=1.0.0